+++ backend/controllers/FaturasController.php
@@ -4,6 +4,7 @@ namespace backend\controllers;
 
 use common\models\Faturas;
 use backend\models\FaturasSearch;
+use Yii;
 use yii\filters\AccessControl;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
@@ -49,13 +50,17 @@ class FaturasController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new FaturasSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
+        if (Yii::$app->user->can('gerirFaturas')) {
+            $searchModel = new FaturasSearch();
+            $dataProvider = $searchModel->search($this->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        }
+
+        return $this->redirect(['index']);
     }
 
     /**
@@ -66,9 +71,13 @@ class FaturasController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('verFaturas')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }
+
+        return $this->redirect(['index']);
     }
 
     /**
@@ -78,19 +87,23 @@ class FaturasController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Faturas();
+        if (Yii::$app->user->can('criarFaturas')) {
+            $model = new Faturas();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
-        } else {
-            $model->loadDefaultValues();
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
         }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        return $this->redirect(['index']);
     }
 
     /**
@@ -102,15 +115,19 @@ class FaturasController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('atualizarFaturas')) {
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
         }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        return $this->redirect(['index']);
     }
 
     /**
@@ -122,13 +139,17 @@ class FaturasController extends Controller
      */
     public function actionDelete($id)
     {
-        $fatura = $this->findModel($id);
-        if ($fatura !== null) {
-            $faturaLinhas = $fatura->getFaturaLinhas()->all();
-            foreach ($faturaLinhas as $linha) {
-                $linha->delete();
+        if (Yii::$app->user->can('apagarFaturas')) {
+            $fatura = $this->findModel($id);
+            if ($fatura !== null) {
+                $faturaLinhas = $fatura->getFaturaLinhas()->all();
+                foreach ($faturaLinhas as $linha) {
+                    $linha->delete();
+                }
+                $fatura->delete();
             }
-            $fatura->delete();
+
+            return $this->redirect(['index']);
         }
 
         return $this->redirect(['index']);
